refactor(data): split out data into 'barebones' vs more 'demo' oriented data
