+++ src/common/data.js
@@ -1,107 +1,3 @@
-import mat4 from 'gl-mat4'
-import bunny from 'bunny'
-
-// /per model data
-let triMat = mat4.identity([]) // create([])
-mat4.translate(triMat, triMat, [1, 10, 20])
-
-const triData = {
-  geometry: {
-    positions: /* [
-      2, 2, 0,
-      1, 2, -2,
-      0, 1, -2
-    ],*/
-    new Float32Array([ // works too!
-      2, 2, 0,
-      1, 2, -2,
-      0, 1, -2
-    ])
-  },
-  transforms: {
-    pos: [0, 0, 0],
-    rot: [0, 0, 0],
-    sca: [1, 1, 1]
-  },
-
-  visuals: {
-    color: [1, 0, 0, 0.5],
-    type: 'mesh'
-  }
-
-}
-
-const bunnyGeom = Object.assign({}, bunny, {id: '0'})
-
-const bunnyData = {
-  id: 'b0',
-  geometry: bunnyGeom, // in truth I do not see cases where bounds would change independantly of geometry, but we might
-  // have bounds WITHOUT geometry
-
-  bounds: {
-    dia: 40,
-    min: [9, 10, 0],
-    max: [15, 10, 4]
-  },
-
-  transforms: {
-    pos: [5, 10, 0],
-    rot: [Math.PI / 2, 0, 0],
-    sca: [1, 1, 1]
-  },
-
-  visuals: {
-    color: [0.02, 0.7, 1, 1], // 07a9ff [1, 1, 0, 0.5],
-    type: 'mesh'
-  },
-
-  meta: {
-    selected: false,
-    pickable: true
-  }
-
-}
-
-const bunnyData2 = {
-  id: 'b1',
-  geometry: bunnyGeom,
-  transforms: {
-    pos: [0, 0, 0],
-    rot: [Math.PI/2, 0, 0],
-    sca: [1, 1, -1]
-  },
-
-  visuals: {
-    color: [0, 1, 1, 0.5],
-    visible: true,
-    type: 'mesh'
-  },
-
-  meta: {
-    selected: false,
-    pickable: true
-  }
-}
-
-const bunnyData3 = {
-  id: 'b2',
-  geometry: bunnyGeom,
-  transforms: {
-    pos: [10, -12.989, 3],
-    rot: [Math.PI/2, 0, 0],
-    sca: [1, 1, 1]
-  },
-  visuals: {
-    color: [1, 1, 0, 0.5],
-    type: 'mesh'
-  },
-
-  meta: {
-    selected: true,
-    pickable: true
-  }
-}
-
 const sceneData = {
   cameras: [
     {
@@ -136,4 +32,4 @@ const sceneData = {
   ]
 }
 
-export { bunnyData, bunnyData2, bunnyData3, sceneData }
+export { sceneData }
